+++ components/TableOfContents/index.tsx
@@ -4,8 +4,8 @@ import Link from 'next/link';
 import { Dispatch, SetStateAction, useEffect, useState } from 'react';
 
 import { TableOfContents } from '@/types/components/data.types';
-import createId from '@/utils/createId';
 import { getIntersectionObserver } from '@/utils/getIntersectionObserver';
+import { createId, getRawString } from '@/utils/getString';
 
 import styles from './styles.module.scss';
 
@@ -34,7 +34,7 @@ const TableOfContentItem = ({
       className={cx('toc-item', level, { current: isActive })}
       onClick={handleClick}
       href={`#${id}`}>
-      {text}
+      {getRawString(text)}
     </Link>
   );
 };
